05_openai-conversational-qa/pyproject.toml
[tool.poetry]
name = "openai-conversational-qa"
version = "0.1.0"
description = ""
authors = ["MuhammadAhsaanAbbasi <mahsaanabbasi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
sqlmodel = "^0.0.16"
psycopg2-binary = "^2.9.9"
python-multipart = "^0.0.9"
types-requests = "^2.32.0.20240602"
orjson = "^3.10.6"
python-dotenv = "^1.0.1"
langchain = "^0.3.0"
langchain-community = "^0.3.0"
langsmith = "^0.1.125"
langserve = "^0.3.0"
langchain-core = "^0.3.5"
langchain-openai = "^0.2.0"
langchainhub = "^0.1.21"
streamlit = "^1.38.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
